requirements.txt
fastapi
uvicorn
discord.py
spotipy
yt-dlp
python-dotenv